requirements.txt
typer
rich
PyGithub
python-dotenv
pyperclip